refactor: move services to hooks directory
