+++ src/components/Trade/TradeInputV2.tsx
@@ -9,8 +9,8 @@ import type { AccountDropdownProps } from 'components/AccountDropdown/AccountDro
 import { SlideTransition } from 'components/SlideTransition'
 import { useSwapper } from 'components/Trade/hooks/useSwapper/useSwapperV2'
 import { getSendMaxAmount } from 'components/Trade/hooks/useSwapper/utils'
-import { useSwapperService } from 'components/Trade/services/useSwapperService'
-import { useTradeAmounts } from 'components/Trade/services/useTradeAmounts'
+import { useSwapperService } from 'components/Trade/hooks/useSwapperService'
+import { useTradeAmounts } from 'components/Trade/hooks/useTradeAmounts'
 import { bnOrZero } from 'lib/bignumber/bignumber'
 import { logger } from 'lib/logger'
 import { selectFeeAssetById } from 'state/slices/assetsSlice/selectors'
