+++ src/components/Trade/TradeRoutes/TradeRoutes.tsx
@@ -4,8 +4,8 @@ import { AnimatePresence } from 'framer-motion'
 import type { RouteComponentProps } from 'react-router-dom'
 import { Redirect, Route, Switch, useLocation } from 'react-router-dom'
 import { Approval } from 'components/Approval/Approval'
+import { useDefaultAssetsService } from 'components/Trade/hooks/useDefaultAssetsService'
 import { SelectAccount } from 'components/Trade/SelectAccount'
-import { useDefaultAssetsService } from 'components/Trade/services/useDefaultAssetsService'
 import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 
 import { useSwapper } from '../hooks/useSwapper/useSwapperV2'
